Adding dependencies in preparation for Heroku deployment
Adding dependencies in preparation for Heroku deployment
as outlined in the following Botpress docs:

https://github.com/botpress/botpress/blob/master/docs/deploy/heroku.md

Added the following lines to `package.json` manually as indicated in the
docs:

```
"engines": {
  "node": ">= 8.0"

```
Also ran `npm install bluebird` which added the following line
to `package.json`:

```
  "bluebird": "^3.5.0",

```

Restarted server using Ctrl+C to shutdown, and `bp start` to restart.
Tested bot on Facebook and the bot was functional and responsive.

Responds to Issue #2.

Next step: `npm install lodash` and test on FB