Added additional dependencies in preparation for Heroku deployment
Added additional dependencies prior to deploying to Heroku.

Ran `npm install lodash`.

Restarted bot server on localhost using Ctrl+C to shut down and `bp start`
to restart. Botpress server restarted cleanly, and bot is responsive
and functional on Facebook.

Responds to Issue #2.

Next steps: See if we need a complete new Heroku instance,
or if we need to make a new one, and if we need to make a new one,
will deleting the existing one mean we can use the same name?

May create new branch prior to additional steps to make a
strong save point at this juncture.