Setting up Heroku deployment
Set up new Heroku instance for the bot at
http://trollbustersbot.herokuapp.com .

Used the 'deploy using Heroku Git' instructions:

Ran `heroku git:remote -a trollbustersbot`

Added Postgres as an add-on

Set env vars for:

- DATABASE
- DATABASE_URL
- BOTPRESS_PASSWORD
- MESSENGER_HOST

These are the .heroku-settings file, which is not committed to
Github for obvious reasons.

Next steps: merge branch into `master` then run `git push heroku master`.
If push is successful, run `heroku open` at the command line. This should
open http://trollbustersbot.herokuapp.com and we should see a Botpress
login screen. Use admin as the user and the pwd we set up in the
environmental variable. At that point, we should be able to see the
Botpress interface. We may need to add the Messenger module, but even
if it is there, we will need to give the module the setup details
available in .facebook-settings (also not committed to Github).
The module needs the App ID, App Secret, App Token, and the hostname
(on localhost we use ngrok, but now that we have a heroku server we
can use trollbustersbot.herokuapp.com as our hostname).

Responds to Issue #2.