Successful deployment of bot to Heroku
Ran `heroku open`, which opened
http://trollbustersbot.herokuapp.com

At that URL, as expected, I saw a Botpress login page. I logged
in using the details in .heroku-settings (this file is not
committed because it contains environmental variables).

Once logged in, I added the Facebook details (App ID, App Token,
App Secret, and Hostname).  I turned off `ngrok` locally (Ctrl +C),
then visited the bot page on Facebook to test that the bot was speaking
to our Heroku instance and not to `ngrok` locally.

Bot was responsive and communicated as expected via Facebook.

Next steps: research how to serve photos (or at least URLs) in
response to Quick Response prompts.

Responds to Issue #3.